# lib/api-zod/src/generated/types/achievement.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Personal Portfolio API
 * OpenAPI spec version: 0.1.0
 */

export interface Achievement {
  id: number;
  title: string;
  description: string;
  /** @nullable */
  date?: string | null;
  /** @nullable */
  icon?: string | null;
  sortOrder: number;
}