# lib/api-zod/src/generated/types/adminSession.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Personal Portfolio API
 * OpenAPI spec version: 0.1.0
 */

export interface AdminSession {
  authenticated: boolean;
  username: string;
}