# lib/api-zod/src/generated/types/certificateInput.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Personal Portfolio API
 * OpenAPI spec version: 0.1.0
 */

export interface CertificateInput {
  name: string;
  issuer: string;
  /** @nullable */
  issueDate?: string | null;
  /** @nullable */
  credentialUrl?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  sortOrder: number;
}