# lib/api-zod/src/generated/types/skillInput.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Personal Portfolio API
 * OpenAPI spec version: 0.1.0
 */

export interface SkillInput {
  name: string;
  category: string;
  level: number;
  /** @nullable */
  icon?: string | null;
  sortOrder: number;
}